refactor: restructure folder into more logical manner
