--- api/graphql.ts
@@ -1,12 +1,0 @@
-import { ApolloServer } from '@apollo/server';
-import { startServerAndCreateNextHandler } from '@as-integrations/next';
-
-import { resolvers } from './../src/graphql/resolver.js';
-import { typeDefs } from './../src/graphql/type-defs.js';
-
-const server = new ApolloServer({
-  resolvers,
-  typeDefs,
-});
-
-export default startServerAndCreateNextHandler(server);
