+++ src/entity/lending.ts
@@ -1,0 +1,3 @@
+import { Entity } from './base';
+
+export interface Lending extends Entity {}
